Rate of Change

An S4 class to store the result of a rate of change estimation.

Slots

year

A numeric vector giving the time points at which the rate fo change is estimated.

estimate

A numeric vector giving the estimation of the rate of change.

calendar

A character string specifying the chronological scale (either "CE", "BP" or "b2k").

hash

A character string giving the 32-byte MD5 hash of the original data file.

Coerce

In the code snippets below, x is an RateOfChange object.

as.data.frame(x)

Coerces to a data.frame.

See also

Author

N. Frerebeau