Layer-Counted Proxy Records Uncertainties

Represents layer-counted proxy records as sequences of probability distributions on absolute, error-free time axes.

Usage

proxy(depth, ...)

# S4 method for numeric
proxy(
  depth,
  proxy,
  proxy_error,
  time,
  time_error,
  from = min(time),
  to = max(time),
  grid = NULL,
  resolution = NULL,
  calendar = c("BP", "CE", "b2k"),
  density = !samples,
  samples = TRUE,
  n = 100
)

# S4 method for ProxyRecord
autoplot(object, ..., raw = FALSE, IQR = TRUE)

# S4 method for ProxyRecord,missing
plot(x, raw = FALSE, IQR = TRUE, ...)

# S4 method for ProxyRecord
summary(object, level = 0.95)

Arguments

depth

A numeric vector giving the depth at which proxy values and calendar ages were measured.

...

Currently not used.

proxy

A numeric vector giving the proxy values.

proxy_error

A numeric vector giving the proxy uncertainties.

time

A numeric vector giving the calendar ages.

time_error

A numeric vector giving the calendar age uncertainties.

from

A length-one numeric vector specifying the starting value of the temporal sequence at which densities are to be estimated.

to

A length-one numeric vector specifying the end value of the temporal sequence at which densities are to be estimated.

grid

A length-one numeric vector specifying the step size (in units of proxy) at which proxy records densities are to be estimated. If NULL (the default), equally spaced points will be used (according to options("chronos.grid")).

resolution

A length-one numeric vector specifying the temporal resolution (in units of time) at which densities are to be estimated. If NULL (the default), the smallest temporal difference between two consecutive records will be used.

calendar

A character string specifying the chronological scale It must be one of "BP" (the default), "CE" or "b2k".

density

A logical scalar: should densities be kept in the results?

samples

A logical scalar: should samples be returned? Defaults to TRUE.

n

An integer specifying the number of item to choose. Only used if samples is TRUE.

raw

A logical scalar: should the raw proxy data be displayed instead of the mean estimates?

IQR

A logical scalar: should the IQR be displayed instead of the 95% confidence intervals?

x, object

A ProxyRecord object.

level

A length-one numeric vector giving the confidence level.

Value

proxy() returns an ProxyRecord object.

autoplot() returns a ggplot object.

plot() is called it for its side-effects: it results in a graphic being displayed (invisibly returns x).

References

Boers, N., Goswami, B. & Ghil, M. (2017). A Complete Representation of Uncertainties in Layer-Counted Paleoclimatic Archives. Climate of the Past, 13(9): 1169-1180. doi:10.5194/cp-13-1169-2017 .

See also

Author

N. Frerebeau

Examples

if (FALSE) {
## Get NGRIP records
data("ngrip2010", package = "folio")
ngrip2010 <- subset(ngrip2010, !is.na(MCE))

## Replicate fig. 3d from Boers et al. (2017)
ngrip_error <- proxy(
  depth = ngrip2010$depth,
  proxy = ngrip2010$delta,
  proxy_error = 0.01,
  time = ngrip2010$age,
  time_error = ngrip2010$MCE,
  grid = 0.01,
  calendar = "b2k"
)

plot(ngrip_error)
}