Create Time Series

Create Time Series

Usage

series(object, scale, ...)

# S4 method for matrix,TimeScale
series(object, scale, start, end = NULL, frequency = 1, names = NULL)

# S4 method for numeric,TimeScale
series(object, scale, start, end = NULL, frequency = 1, names = NULL)

# S4 method for data.frame,TimeScale
series(object, scale, start, end = NULL, frequency = 1, names = NULL)

Arguments

object

A numeric vector or matrix of the observed time-series values. A data.frame will be coerced to a numeric matrix via data.matrix().

scale

A TimeScale object (see era()).

...

Currently not used.

start

A length-one numeric vector specifying the year of the first observation.

end

A length-one numeric vector specifying the year of the last observation.

frequency

A length-one numeric vector specifying the time difference between two observations (resolution), in years.

names

A character string specifying the names of the time series.

Value

A TimeSeries object.

Details

series() only supports data sampled at equidistant points in time, expressed in decimal years (1950 means 1950.0, i.e. the beginning of the year 1950).

See also

Author

N. Frerebeau

Examples

## Univariate
z <- series(rnorm(300), era("BCE"), start = 1000)

names(z)
#> [1] "S1"
start(z)
#> [1] 1000
end(z)
#> [1] 701
time(z)
#>   [1] 1000  999  998  997  996  995  994  993  992  991  990  989  988  987  986
#>  [16]  985  984  983  982  981  980  979  978  977  976  975  974  973  972  971
#>  [31]  970  969  968  967  966  965  964  963  962  961  960  959  958  957  956
#>  [46]  955  954  953  952  951  950  949  948  947  946  945  944  943  942  941
#>  [61]  940  939  938  937  936  935  934  933  932  931  930  929  928  927  926
#>  [76]  925  924  923  922  921  920  919  918  917  916  915  914  913  912  911
#>  [91]  910  909  908  907  906  905  904  903  902  901  900  899  898  897  896
#> [106]  895  894  893  892  891  890  889  888  887  886  885  884  883  882  881
#> [121]  880  879  878  877  876  875  874  873  872  871  870  869  868  867  866
#> [136]  865  864  863  862  861  860  859  858  857  856  855  854  853  852  851
#> [151]  850  849  848  847  846  845  844  843  842  841  840  839  838  837  836
#> [166]  835  834  833  832  831  830  829  828  827  826  825  824  823  822  821
#> [181]  820  819  818  817  816  815  814  813  812  811  810  809  808  807  806
#> [196]  805  804  803  802  801  800  799  798  797  796  795  794  793  792  791
#> [211]  790  789  788  787  786  785  784  783  782  781  780  779  778  777  776
#> [226]  775  774  773  772  771  770  769  768  767  766  765  764  763  762  761
#> [241]  760  759  758  757  756  755  754  753  752  751  750  749  748  747  746
#> [256]  745  744  743  742  741  740  739  738  737  736  735  734  733  732  731
#> [271]  730  729  728  727  726  725  724  723  722  721  720  719  718  717  716
#> [286]  715  714  713  712  711  710  709  708  707  706  705  704  703  702  701

## Multivariate
z <- series(matrix(rnorm(300), 100, 3), era("CE"), start = 1000)

names(z)
#> [1] "S1" "S2" "S3"
start(z)
#> [1] 1000
end(z)
#> [1] 1099
time(z)
#>   [1] 1000 1001 1002 1003 1004 1005 1006 1007 1008 1009 1010 1011 1012 1013 1014
#>  [16] 1015 1016 1017 1018 1019 1020 1021 1022 1023 1024 1025 1026 1027 1028 1029
#>  [31] 1030 1031 1032 1033 1034 1035 1036 1037 1038 1039 1040 1041 1042 1043 1044
#>  [46] 1045 1046 1047 1048 1049 1050 1051 1052 1053 1054 1055 1056 1057 1058 1059
#>  [61] 1060 1061 1062 1063 1064 1065 1066 1067 1068 1069 1070 1071 1072 1073 1074
#>  [76] 1075 1076 1077 1078 1079 1080 1081 1082 1083 1084 1085 1086 1087 1088 1089
#>  [91] 1090 1091 1092 1093 1094 1095 1096 1097 1098 1099